+++ lib/google_my_business.dart
@@ -1,34 +1,31 @@
 /// Flutter package to work with Google My Business API.
-/// Simplifies connection with GMB API, contains wrappers around requests & responses for multiple endpoints including location, review and others.
-///  More info can be found at https://developers.google.com/my-business
+/// Simplifies communication & interaction with GMB API endpoints.
 library google_my_business;
 
-/// Review
-export "src/models/review/review.dart";
-export "src/models/review/reviews.dart";
-export "src/models/review/reviewer.dart";
-export "src/models/review/review_reply.dart";
-
-/// Location
-export "src/models/location/location.dart";
-export "src/models/location/postal_address.dart";
+export "src/gmb_api.dart" show GMBAPI;
+export 'src/managers/locations_manager.dart' show LocationsManager;
+export 'src/managers/reviews_manager.dart' show ReviewsManager;
 export "src/models/location/category/category.dart";
+export "src/models/location/category/more_hours_type.dart";
 export "src/models/location/category/service_type.dart";
 export "src/models/location/common/date.dart";
 export "src/models/location/common/latlng.dart";
 export "src/models/location/hours/business_hours.dart";
 export "src/models/location/hours/day_of_week.dart";
-export "src/models/location/hours/special_hours.dart";
+export "src/models/location/hours/more_hours.dart";
 export "src/models/location/hours/special_hour_period.dart";
+export "src/models/location/hours/special_hours.dart";
 export "src/models/location/hours/time_period.dart";
+export "src/models/location/location.dart";
 export "src/models/location/metadata/ad_words_location_extensions.dart";
 export "src/models/location/metadata/location_key.dart";
 export "src/models/location/metadata/location_state.dart";
 export "src/models/location/metadata/metadata.dart";
 export "src/models/location/metadata/profile.dart";
 export "src/models/location/metadata/relationship_data.dart";
-export "src/models/location/open_info/open_info.dart";
 export "src/models/location/open_info/open_for_business.dart";
+export "src/models/location/open_info/open_info.dart";
+export "src/models/location/postal_address.dart";
 export "src/models/location/price_list/item.dart";
 export "src/models/location/price_list/label.dart";
 export "src/models/location/price_list/money.dart";
@@ -39,6 +36,9 @@ export "src/models/location/service_area/place_info.dart";
 export "src/models/location/service_area/places.dart";
 export "src/models/location/service_area/point_radius.dart";
 export "src/models/location/service_area/service_area_business.dart";
-
-export "src/gmb_api.dart" show GMBAPI;
-export "src/locations_manager.dart" show LocationsManager;
+export "src/models/review/review.dart";
+export "src/models/review/review_reply.dart";
+export "src/models/review/reviewer.dart";
+export "src/models/review/reviews.dart";
+export "src/models/review/star_rating.dart";
+export "src/models/error.dart";