+++ lib/src/models/review/star_rating.dart
@@ -1,0 +1,2 @@
+/// The star rating out of five, where five is the highest rated.
+enum StarRating { STAR_RATING_UNSPECIFIED, ONE, TWO, THREE, FOUR, FIVE }